Add an options block to the JSLINT() call
indent - This was to prevent false indent warnings, where jslint wanted
everything at column 1.

sloppy - do not require the 'use strict' directive.

maxerr - increase the maximum number of errors shown from 50 to 100.

predef - Made 'jQuery' be predefined, figured this would be benign when
not using jQuery, but very helpful if you are.